o05_12クラスの修正

+++ src/main/java/ruly/lecture/ooka/practice/oop/o05_strategy/expert/o05_12/PlemiumUser.java
@@ -3,7 +3,7 @@ package ruly.lecture.ooka.practice.oop.o05_strategy.expert.o05_12;
 import java.util.List;
 
 public class PlemiumUser extends UserAccount {
-	public PlemiumUser(String userId, int currencyAmount, String currencyCode, List<String> ipList,
+	public PlemiumUser(String userId, double currencyAmount, String currencyCode, List<String> ipList,
 			CurrencyExchangeStrategy currenctExchangeStrategy, SecurityStrategy securityStrategy) {
 		super(userId, currencyAmount, currencyCode, ipList, currenctExchangeStrategy, securityStrategy);
 	}
